move and rename text input to make it reusable
