+++ src/components/TextInputWithAssetSelector/package.json
@@ -1,0 +1,7 @@
+{
+    "name": "TextInputWithAssetSelector.js",
+    "version": "0.0.0",
+    "private": true,
+    "main": "./TextInputWithAssetSelector.js"
+  }
+
